# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: XBoard
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
paths:
...
  /api/v1/server/TrojanTidalab/config:
    get:
      summary: ''
      operationId: getApiV1ServerTrojanTidalabConfig
      description: ''
      parameters: []
      responses:
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: validation.required
                  errors:
                    token:
                      - validation.required
                properties:
                  message:
                    type: string
                    example: validation.required
                  errors:
                    type: object
                    properties:
                      token:
                        type: array
                        example:
                          - validation.required
                        items:
                          type: string
      tags:
        - '服务端端点 API (Server Nodes)'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                node_id:
                  type: string
                  description: ''
                  example: vel
                local_port:
                  type: string
                  description: ''
                  example: qui
              required:
                - node_id
                - local_port
      security: []
...
  /api/v1/client/subscribe:
    get:
      summary: '获取节点订阅配置 (Subscribe)'
      operationId: Subscribe
      description: "各大代理客户端（如Clash, V2ray, Shadowrocket等）使用的统一订阅获取接口。\n根据请求头的 User-Agent 或 flag 动态下发对应格式的协议节点配置。"
      parameters:
        -
          in: query
          name: flag
          description: '强制指定客户端类型标识(如 clash, v2ray, shadowrocket)'
          example: error
          required: false
          schema:
            type: string
            description: '强制指定客户端类型标识(如 clash, v2ray, shadowrocket)'
            example: error
        -
          in: query
          name: types
          description: '取特定的节点类型（如 vmess|vless|trojan)'
          example: quas
          required: false
          schema:
            type: string
            description: '取特定的节点类型（如 vmess|vless|trojan)'
            example: quas
        -
          in: query
          name: filter
          description: 只保留包含特定关键字名称的节点
          example: ex
          required: false
          schema:
            type: string
            description: 只保留包含特定关键字名称的节点
            example: ex
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: plain/text
      tags:
        - '用户端 API (User & Guest)'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                types:
                  type: string
                  description: ''
                  example: aut
                  nullable: true
                filter:
                  type: string
                  description: ''
                  example: quas
                  nullable: true
                flag:
                  type: string
                  description: ''
                  example: nostrum
                  nullable: true
      security: []
...
  /api/v1/guest/telegram/webhook:
    post:
      summary: 'Telegram Bot Webhook 回调地址'
      operationId: telegramBotWebhook
      description: "用于接收 Telegram 服务器端推送过来的用户交互消息、加群请求等，通过 Hook 机制分发处理逻辑。\n该地址高度保密且需携带设定好的 accessToken 防止被恶意请求。"
      parameters: []
      responses: {  }
      tags:
        - '用户端 API (User & Guest)'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                access_token:
                  type: string
                  description: Telegram通信密钥配置
                  example: md5string
              required:
                - access_token
      security: []
  '/api/v1/guest/payment/notify/{method}/{uuid}':
    get:
      summary: '支付网关异步回调通知 (Webhook)'
      operationId: Webhook
      description: "用于接收第三方支付平台异步发送的订单支付结果通知，验证签名后更改内部订单状态并下发服务。\n该接口无需身份认证，完全依赖网关签名校验。"
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: plain/text
      tags:
        - '用户端 API (User & Guest)'
      security: []
    parameters:
      -
        in: path
        name: method
        description: ''
        example: natus
        required: true
        schema:
          type: string
      -
        in: path
        name: uuid
        description: ''
        example: 145d25e7-6a57-3fa8-b247-30ebcaf7daa8
        required: true
        schema:
          type: string
...
  /api/v2/57a7189f/plugin/config:
    get:
      summary: 获取某个插件运行选项配置
      operationId: ''
      description: ''
      parameters:
        -
          in: query
          name: code
          description: 对应参数的插件包名
          example: nostrum
          required: true
          schema:
            type: string
            description: 对应参数的插件包名
            example: nostrum
      responses:
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthorized
                properties:
                  message:
                    type: string
                    example: Unauthorized
      tags:
        - '管理后台 API (Admin)'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: ''
                  example: distinctio
              required:
                - code
      security: []
    post:
      summary: 提交保存插件设置
      operationId: ''
      description: 对插件自己抛出定义的需要表单配置执行设值。
      parameters: []
      responses: {  }
      tags:
        - '管理后台 API (Admin)'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: 配置目标关联包名
                  example: beatae
                config:
                  type: array
                  description: 配置键值对
                  example:
                    - enim
                  items:
                    type: string
              required:
                - code
                - config
      security: []
...